requirements.txt
rendercv[full]
jinja2 >= 3.1.6
pyyaml >= 6.0.2
